pyproject.toml
[project]
name = "fablestar"
version = "0.1.0"
description = "Vibe-coded MUD platform with LLM integration and hot-reload."
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "redis[hiredis]>=5.0.0",
    "asyncpg>=0.29.0",
    "pydantic[email]>=2.5.0",
    "jinja2>=3.1.0",
    "watchdog>=3.0.0",
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "bcrypt>=4.1.0",
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "alembic>=1.13.0",
    "python-multipart>=0.0.9",
    "httpx>=0.26.0",
    "openai>=1.12.0",
    "psutil>=5.9.0",
    "PyYAML>=6.0.0",
    "SQLAlchemy>=2.0.0",
]

[project.scripts]
fablestar = "fablestar.__main__:main"
scaffold = "cli.scaffold:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true
